// src/lib/proto/RCRSLogProto.ts
export interface LogProto {
  start?: StartLogProto | undefined;
  initialCondition?: InitialConditionsLogProto | undefined;
  command?: CommandLogProto | undefined;
  perception?: PerceptionLogProto | undefined;
  config?: ConfigLogProto | undefined;
  update?: UpdatesLogProto | undefined;
  end?: EndLogProto | undefined;
}

export interface StartLogProto {
}

export interface InitialConditionsLogProto {
  entities: EntityProto[];
}

export interface CommandLogProto {
  time: number;
  commands: MessageProto[];
}

export interface PerceptionLogProto {
  time: number;
  entityID: number;
  visible?: ChangeSetProto | undefined;
  communications: MessageProto[];
}

export interface ConfigLogProto {
  config?: ConfigProto | undefined;
}

export interface UpdatesLogProto {
  time: number;
  changes?: ChangeSetProto | undefined;
}

export interface EndLogProto {
}

function createBaseLogProto(): LogProto {
  return {
    start: undefined,
    initialCondition: undefined,
    command: undefined,
    perception: undefined,
    config: undefined,
    update: undefined,
    end: undefined,
  };
}

export const LogProto: MessageFns<LogProto> = {
  encode(message: LogProto, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.start !== undefined) {
      StartLogProto.encode(message.start, writer.uint32(10).fork()).join();
    }
    if (message.initialCondition !== undefined) {
      InitialConditionsLogProto.encode(message.initialCondition, writer.uint32(18).fork()).join();
    }
    if (message.command !== undefined) {
      CommandLogProto.encode(message.command, writer.uint32(26).fork()).join();
    }
    if (message.perception !== undefined) {
      PerceptionLogProto.encode(message.perception, writer.uint32(34).fork()).join();
    }
    if (message.config !== undefined) {
      ConfigLogProto.encode(message.config, writer.uint32(42).fork()).join();
    }
    if (message.update !== undefined) {
      UpdatesLogProto.encode(message.update, writer.uint32(50).fork()).join();
    }
    if (message.end !== undefined) {
      EndLogProto.encode(message.end, writer.uint32(58).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): LogProto {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseLogProto();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.start = StartLogProto.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.initialCondition = InitialConditionsLogProto.decode(reader, reader.uint32());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.command = CommandLogProto.decode(reader, reader.uint32());
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.perception = PerceptionLogProto.decode(reader, reader.uint32());
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.config = ConfigLogProto.decode(reader, reader.uint32());
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.update = UpdatesLogProto.decode(reader, reader.uint32());
          continue;
        }
        case 7: {
          if (tag !== 58) {
            break;
          }

          message.end = EndLogProto.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<LogProto>, I>>(base?: I): LogProto {
    return LogProto.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<LogProto>, I>>(object: I): LogProto {
    const message = createBaseLogProto();
    message.start = (object.start !== undefined && object.start !== null)
      ? StartLogProto.fromPartial(object.start)
      : undefined;
    message.initialCondition = (object.initialCondition !== undefined && object.initialCondition !== null)
      ? InitialConditionsLogProto.fromPartial(object.initialCondition)
      : undefined;
    message.command = (object.command !== undefined && object.command !== null)
      ? CommandLogProto.fromPartial(object.command)
      : undefined;
    message.perception = (object.perception !== undefined && object.perception !== null)
      ? PerceptionLogProto.fromPartial(object.perception)
      : undefined;
    message.config = (object.config !== undefined && object.config !== null)
      ? ConfigLogProto.fromPartial(object.config)
      : undefined;
    message.update = (object.update !== undefined && object.update !== null)
      ? UpdatesLogProto.fromPartial(object.update)
      : undefined;
    message.end = (object.end !== undefined && object.end !== null) ? EndLogProto.fromPartial(object.end) : undefined;
    return message;
  },
};

function createBaseStartLogProto(): StartLogProto {
  return {};
}

export const StartLogProto: MessageFns<StartLogProto> = {
  encode(_: StartLogProto, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): StartLogProto {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseStartLogProto();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<StartLogProto>, I>>(base?: I): StartLogProto {
    return StartLogProto.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<StartLogProto>, I>>(_: I): StartLogProto {
    const message = createBaseStartLogProto();
    return message;
  },
};

function createBaseInitialConditionsLogProto(): InitialConditionsLogProto {
  return { entities: [] };
}

export const InitialConditionsLogProto: MessageFns<InitialConditionsLogProto> = {
  encode(message: InitialConditionsLogProto, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.entities) {
      EntityProto.encode(v!, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): InitialConditionsLogProto {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseInitialConditionsLogProto();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.entities.push(EntityProto.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<InitialConditionsLogProto>, I>>(base?: I): InitialConditionsLogProto {
    return InitialConditionsLogProto.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<InitialConditionsLogProto>, I>>(object: I): InitialConditionsLogProto {
    const message = createBaseInitialConditionsLogProto();
    message.entities = object.entities?.map((e) => EntityProto.fromPartial(e)) || [];
    return message;
  },
};

function createBaseCommandLogProto(): CommandLogProto {
  return { time: 0, commands: [] };
}

export const CommandLogProto: MessageFns<CommandLogProto> = {
  encode(message: CommandLogProto, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.time !== 0) {
      writer.uint32(8).int32(message.time);
    }
    for (const v of message.commands) {
      MessageProto.encode(v!, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CommandLogProto {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCommandLogProto();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.time = reader.int32();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.commands.push(MessageProto.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<CommandLogProto>, I>>(base?: I): CommandLogProto {
    return CommandLogProto.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<CommandLogProto>, I>>(object: I): CommandLogProto {
    const message = createBaseCommandLogProto();
    message.time = object.time ?? 0;
    message.commands = object.commands?.map((e) => MessageProto.fromPartial(e)) || [];
    return message;
  },
};

function createBasePerceptionLogProto(): PerceptionLogProto {
  return { time: 0, entityID: 0, visible: undefined, communications: [] };
}

export const PerceptionLogProto: MessageFns<PerceptionLogProto> = {
  encode(message: PerceptionLogProto, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.time !== 0) {
      writer.uint32(8).int32(message.time);
    }
    if (message.entityID !== 0) {
      writer.uint32(16).int32(message.entityID);
    }
    if (message.visible !== undefined) {
      ChangeSetProto.encode(message.visible, writer.uint32(26).fork()).join();
    }
    for (const v of message.communications) {
      MessageProto.encode(v!, writer.uint32(34).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): PerceptionLogProto {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasePerceptionLogProto();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.time = reader.int32();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.entityID = reader.int32();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.visible = ChangeSetProto.decode(reader, reader.uint32());
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.communications.push(MessageProto.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<PerceptionLogProto>, I>>(base?: I): PerceptionLogProto {
    return PerceptionLogProto.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<PerceptionLogProto>, I>>(object: I): PerceptionLogProto {
    const message = createBasePerceptionLogProto();
    message.time = object.time ?? 0;
    message.entityID = object.entityID ?? 0;
    message.visible = (object.visible !== undefined && object.visible !== null)
      ? ChangeSetProto.fromPartial(object.visible)
      : undefined;
    message.communications = object.communications?.map((e) => MessageProto.fromPartial(e)) || [];
    return message;
  },
};
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}